+++ src/components/App/TitleHeader.js
@@ -2,11 +2,12 @@ import React, {useState, useEffect, useContext} from 'react';
 import Typography from '@material-ui/core/Typography';
 import ForwardIcon from '@material-ui/icons/FastForward';
 import Button from '@material-ui/core/Button';
-import bg from '../../assets/images/bg-new.png';
-import OptionsButton from '../OptionsButton';
 import {UserContext} from '../../store/context/userContext';
+import {PRISMIC_TITLE_BG} from '../../utils/constants';
+import {useCmsAsset} from '../../hooks/api-hooks'
 
 const TitleHeader = (props) => {
+    const bgimg = useCmsAsset(PRISMIC_TITLE_BG);
     const {country:country_alt, region:region_alt, place: place_alt, referrer} = props;
     const [state, setState] = React.useState({activeStep : 0, place: place_alt, country : country_alt, region: region_alt})
     const [ffopen, setFfopen] = useState(false);
@@ -24,7 +25,7 @@ const TitleHeader = (props) => {
     return (
         <>
         <Typography color="textPrimary" variant="h1" component="h1" align="left" 
-        style={{backgroundImage:`url(${bg})`, backgroundRepeat:'no-repeat',backgroundPosition: 'right top', backgroundSize: 'auto 100%', backgroundColor: '#0c39e3', fontWeight:'bold', fontSize:'4rem', padding:'24px', color: 'rgba(255, 255, 255, 0.7)', marginBottom:0}} gutterBottom>
+        style={{backgroundImage:`url(${bgimg})`, backgroundRepeat:'no-repeat',backgroundPosition: 'right top', backgroundSize: 'auto 100%', backgroundColor: '#0c39e3', fontWeight:'bold', fontSize:'4rem', padding:'24px', color: 'rgba(255, 255, 255, 0.7)', marginBottom:0}} gutterBottom>
                     Know Your Prayer times {ffopen && <ForwardIcon className="landing-navs" onClick={handleFF} fontSize="large" style={{color:'#fff', fontSize: '2.8rem', top: '5px', position: 'relative'}} />}
         </Typography>